publish message to queue to invoke lambda
